refactor(j-s): address review comments on rich text editor
- Move highlight color picker styles into their own file
- Stabilize editorProps passed to useEditor
